backend/requirements.txt
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
certifi==2024.7.4
cffi==1.17.0
channels==4.1.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==43.0.0
daphne==4.1.2
defusedxml==0.8.0rc2
Django==4.2.15
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
hyperlink==21.0.0
idna==3.7
incremental==24.7.2
oauthlib==3.2.2
pillow==10.4.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==24.2.1
pyotp==2.9.0
pypng==0.20220715.0
python3-openid==3.2.0
qrcode==7.4.2
requests==2.32.3
requests-oauthlib==2.0.0
service-identity==24.1.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
tomli==2.0.1
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
urllib3==2.2.2
zope.interface==7.0.3
psycopg
psycopg2
psycopg2-binary